Move custom click functions from common.py to group_options.py
